MGDOBR-1294: Refactor TemplateProvider in v2
