+++ shard-operator-parent/shard-operator-core/src/main/java/com/redhat/service/smartevents/shard/operator/core/providers/TemplateProvider.java
@@ -10,26 +10,10 @@ import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.api.model.Secret;
-import io.fabric8.kubernetes.api.model.Service;
-import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.fabric8.kubernetes.api.model.networking.v1.Ingress;
-import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.openshift.api.model.Route;
-import io.fabric8.openshift.api.model.monitoring.v1.ServiceMonitor;
 
 public interface TemplateProvider {
-    Secret loadBridgeExecutorSecretTemplate(HasMetadata resource, TemplateImportConfig config);
-
-    Deployment loadBridgeExecutorDeploymentTemplate(HasMetadata resource, TemplateImportConfig config);
-
-    Service loadBridgeExecutorServiceTemplate(HasMetadata resource, TemplateImportConfig config);
-
-    Route loadBridgeIngressOpenshiftRouteTemplate(HasMetadata resource, TemplateImportConfig config);
-
-    Ingress loadBridgeIngressKubernetesIngressTemplate(HasMetadata resource, TemplateImportConfig config);
-
-    ServiceMonitor loadServiceMonitorTemplate(CustomResource resource, TemplateImportConfig config);
-
     Secret loadBridgeIngressSecretTemplate(HasMetadata resource, TemplateImportConfig config);
 
     ConfigMap loadBridgeIngressConfigMapTemplate(HasMetadata resource, TemplateImportConfig config);
@@ -37,6 +21,10 @@ public interface TemplateProvider {
     KnativeBroker loadBridgeIngressBrokerTemplate(HasMetadata resource, TemplateImportConfig config);
 
     AuthorizationPolicy loadBridgeIngressAuthorizationPolicyTemplate(HasMetadata resource, TemplateImportConfig config);
+
+    Route loadBridgeIngressOpenshiftRouteTemplate(HasMetadata resource, TemplateImportConfig config);
+
+    Ingress loadBridgeIngressKubernetesIngressTemplate(HasMetadata resource, TemplateImportConfig config);
 
     Secret loadObservabilitySecretTemplate(String name, String namespace);
