+++ shard-operator-parent/shard-operator-core/src/main/java/com/redhat/service/smartevents/shard/operator/core/providers/TemplateProviderImpl.java
@@ -17,84 +17,31 @@ import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.api.model.Secret;
-import io.fabric8.kubernetes.api.model.Service;
-import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.fabric8.kubernetes.api.model.networking.v1.Ingress;
-import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.client.utils.Serialization;
 import io.fabric8.openshift.api.model.Route;
-import io.fabric8.openshift.api.model.monitoring.v1.ServiceMonitor;
 
 @ApplicationScoped
 public class TemplateProviderImpl implements TemplateProvider {
 
     private static final String TEMPLATES_DIR = "/templates";
-    private static final String BRIDGE_INGRESS_OPENSHIFT_ROUTE_PATH = TEMPLATES_DIR + "/bridge-ingress-openshift-route.yaml";
-    private static final String BRIDGE_INGRESS_KUBERNETES_INGRESS_PATH = TEMPLATES_DIR + "/bridge-ingress-kubernetes-ingress.yaml";
     private static final String BRIDGE_INGRESS_SECRET_PATH = TEMPLATES_DIR + "/bridge-ingress-secret.yaml";
     private static final String BRIDGE_INGRESS_CONFIGMAP_PATH = TEMPLATES_DIR + "/bridge-ingress-broker-configmap.yaml";
     private static final String BRIDGE_INGRESS_BROKER_PATH = TEMPLATES_DIR + "/bridge-ingress-broker.yaml";
     private static final String BRIDGE_INGRESS_AUTHORIZATION_POLICY_PATH = TEMPLATES_DIR + "/bridge-ingress-authorization-policy.yaml";
-
-    private static final String BRIDGE_EXECUTOR_DEPLOYMENT_PATH = TEMPLATES_DIR + "/bridge-executor-deployment.yaml";
-    private static final String BRIDGE_EXECUTOR_SERVICE_PATH = TEMPLATES_DIR + "/bridge-executor-service.yaml";
-    private static final String BRIDGE_EXECUTOR_SECRET_PATH = TEMPLATES_DIR + "/bridge-executor-secret.yaml";
+    private static final String BRIDGE_INGRESS_OPENSHIFT_ROUTE_PATH = TEMPLATES_DIR + "/bridge-ingress-openshift-route.yaml";
+    private static final String BRIDGE_INGRESS_KUBERNETES_INGRESS_PATH = TEMPLATES_DIR + "/bridge-ingress-kubernetes-ingress.yaml";
 
     private static final String OBSERVABILITY_SECRET_PATH = TEMPLATES_DIR + "/observability-secret.yaml";
 
-    private static final String SERVICE_MONITOR_PATH = TEMPLATES_DIR + "/service-monitor.yaml";
-
     private static final String ISTIO_GATEWAY_PATH = TEMPLATES_DIR + "/gateway.yaml";
-
     private static final String ISTIO_VIRTUAL_SERVICE_PATH = TEMPLATES_DIR + "/virtual-service-kafka-broker.yaml";
 
     private static final String JWT_REQUEST_AUTHENTICATION_PATH = TEMPLATES_DIR + "/jwt-request-authentication.yaml";
 
     @Override
-    public Deployment loadBridgeExecutorDeploymentTemplate(HasMetadata resource, TemplateImportConfig config) {
-        Deployment deployment = loadYaml(Deployment.class, BRIDGE_EXECUTOR_DEPLOYMENT_PATH);
-        updateMetadata(resource, deployment.getMetadata(), config);
-        return deployment;
-    }
-
-    @Override
-    public Service loadBridgeExecutorServiceTemplate(HasMetadata resource, TemplateImportConfig config) {
-        Service service = loadYaml(Service.class, BRIDGE_EXECUTOR_SERVICE_PATH);
-        updateMetadata(resource, service.getMetadata(), config);
-        return service;
-    }
-
-    @Override
-    public Route loadBridgeIngressOpenshiftRouteTemplate(HasMetadata resource, TemplateImportConfig config) {
-        Route route = loadYaml(Route.class, BRIDGE_INGRESS_OPENSHIFT_ROUTE_PATH);
-        updateMetadata(resource, route.getMetadata(), config);
-        return route;
-    }
-
-    @Override
-    public Ingress loadBridgeIngressKubernetesIngressTemplate(HasMetadata resource, TemplateImportConfig config) {
-        Ingress ingress = loadYaml(Ingress.class, BRIDGE_INGRESS_KUBERNETES_INGRESS_PATH);
-        updateMetadata(resource, ingress.getMetadata(), config);
-        return ingress;
-    }
-
-    @Override
-    public ServiceMonitor loadServiceMonitorTemplate(CustomResource resource, TemplateImportConfig config) {
-        final ServiceMonitor serviceMonitor = loadYaml(ServiceMonitor.class, SERVICE_MONITOR_PATH);
-        updateMetadata(resource, serviceMonitor.getMetadata(), config);
-        return serviceMonitor;
-    }
-
-    @Override
     public Secret loadBridgeIngressSecretTemplate(HasMetadata resource, TemplateImportConfig config) {
         final Secret secret = loadYaml(Secret.class, BRIDGE_INGRESS_SECRET_PATH);
-        updateMetadata(resource, secret.getMetadata(), config);
-        return secret;
-    }
-
-    @Override
-    public Secret loadBridgeExecutorSecretTemplate(HasMetadata resource, TemplateImportConfig config) {
-        final Secret secret = loadYaml(Secret.class, BRIDGE_EXECUTOR_SECRET_PATH);
         updateMetadata(resource, secret.getMetadata(), config);
         return secret;
     }
@@ -118,6 +65,20 @@ public class TemplateProviderImpl implements TemplateProvider {
         AuthorizationPolicy authorizationPolicy = loadYaml(AuthorizationPolicy.class, BRIDGE_INGRESS_AUTHORIZATION_POLICY_PATH);
         updateMetadata(resource, authorizationPolicy.getMetadata(), config);
         return authorizationPolicy;
+    }
+
+    @Override
+    public Route loadBridgeIngressOpenshiftRouteTemplate(HasMetadata resource, TemplateImportConfig config) {
+        Route route = loadYaml(Route.class, BRIDGE_INGRESS_OPENSHIFT_ROUTE_PATH);
+        updateMetadata(resource, route.getMetadata(), config);
+        return route;
+    }
+
+    @Override
+    public Ingress loadBridgeIngressKubernetesIngressTemplate(HasMetadata resource, TemplateImportConfig config) {
+        Ingress ingress = loadYaml(Ingress.class, BRIDGE_INGRESS_KUBERNETES_INGRESS_PATH);
+        updateMetadata(resource, ingress.getMetadata(), config);
+        return ingress;
     }
 
     @Override
