+++ shard-operator-parent/shard-operator-v1/src/main/resources/templates/bridge-executor-service.yaml
