+++ shard-operator-parent/shard-operator-v1/src/main/resources/templates/service-monitor.yaml
