+++ shard-operator-parent/shard-operator-v2/src/main/java/com/redhat/service/smartevents/shard/operator/v2/providers/TemplateProviderV2.java
@@ -1,10 +1,11 @@
 package com.redhat.service.smartevents.shard.operator.v2.providers;
 
 import com.redhat.service.smartevents.shard.operator.core.providers.TemplateImportConfig;
+import com.redhat.service.smartevents.shard.operator.core.providers.TemplateProvider;
 import com.redhat.service.smartevents.shard.operator.v2.resources.CamelIntegration;
 
 import io.fabric8.kubernetes.client.CustomResource;
 
-public interface TemplateProviderV2 {
+public interface TemplateProviderV2 extends TemplateProvider {
     CamelIntegration loadCamelIntegrationTemplate(CustomResource owner, TemplateImportConfig config);
 }
